Consolidate co-emergence utilities
