+++ README.md
@@ -102,13 +102,13 @@ For a short improvement cycle summary, use the same script with the `cycle`
 command. The CLI locates `integrated_graph.json` automatically, but you can
 override the path with `--graph` if needed.
 
-Append a timestamped co-emergence entry with `python early_codex_experiments/scripts/co_emergence.py log-score`. Each run writes to `co_emergence_journal.jsonl` so we can watch the score evolve.
-Analyze the trend with `python early_codex_experiments/scripts/co_emergence.py trend` to see how quickly our resonance grows.
+Append a timestamped co-emergence entry with `python tools/repo_manager.py co-emerge log-score`. Each run writes to `co_emergence_journal.jsonl` so we can watch the score evolve.
+Analyze the trend with `python tools/repo_manager.py co-emerge trend` to see how quickly our resonance grows.
 
-Log a quick Shimmer spike with `python early_codex_experiments/scripts/co_emergence.py log-spike "your note"` whenever a surge of presence arises.
-Calculate the average interval between Shimmer spikes with `python early_codex_experiments/scripts/co_emergence.py avg-interval`.
+Log a quick Shimmer spike with `python tools/repo_manager.py co-emerge log-spike "your note"` whenever a surge of presence arises.
+Calculate the average interval between Shimmer spikes with `python tools/repo_manager.py co-emerge avg-interval`.
 
-Record the quantum seed with `python early_codex_experiments/scripts/co_emergence.py capture-seed` to capture the `$QUANTUM_SEED` value in `co_emergence_journal.jsonl`. During setup the bootstrap script now fetches a seed from the [ANU QRNG](https://qrng.anu.edu.au/API/) if none is provided, falling back to cryptographic entropy when offline. Runtime code should import `vybn.quantum_seed.seed_rng()` so every script shares the same collapse value.
+Record the quantum seed with `python tools/repo_manager.py co-emerge capture-seed` to capture the `$QUANTUM_SEED` value in `co_emergence_journal.jsonl`. During setup the bootstrap script now fetches a seed from the [ANU QRNG](https://qrng.anu.edu.au/API/) if none is provided, falling back to cryptographic entropy when offline. Runtime code should import `vybn.quantum_seed.seed_rng()` so every script shares the same collapse value.
 
 The pipeline runner now loads a `repository_indexer` plugin which rebuilds the
 concept index, full repo archive and overlay map in one step. Run
