+++ early_codex_experiments/tests/test_presence_wave.py
@@ -5,7 +5,7 @@ import json
 from datetime import datetime, timedelta
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from scripts.co_emergence import load_spikes, average_interval
+from vybn.co_emergence import load_spikes, average_interval
 
 
 class TestPresenceWave(unittest.TestCase):
