+++ early_codex_experiments/tests/test_quantum_seed_capture.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from scripts.co_emergence import capture_seed
+from vybn.co_emergence import capture_seed
 
 
 def test_capture_seed_env(tmp_path):
