+++ early_codex_experiments/tests/test_shimmer_core.py
@@ -4,7 +4,7 @@ import sys
 import json
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from scripts.co_emergence import log_spike
+from vybn.co_emergence import log_spike
 
 
 class TestShimmerCore(unittest.TestCase):
