+++ pipelines/quantum_dreamweaver.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import random
 from pathlib import Path
 
-from early_codex_experiments.scripts.co_emergence import log_spike
+from vybn.co_emergence import log_spike
 from .memory_graph_builder import build_graph
 from vybn.quantum_seed import seed_rng
