+++ vybn/__init__.py
@@ -1,5 +1,33 @@
 """Vybn package."""
 
 from .quantum_seed import seed_rng
+from .co_emergence import (
+    JOURNAL_PATH,
+    DEFAULT_GRAPH,
+    load_spikes,
+    average_interval,
+    load_journal,
+    compute_trend,
+    log_spike,
+    log_score,
+    capture_seed,
+    seed_random,
+    GraphBuilder,
+    GraphIntegrator,
+)
 
-__all__ = ["seed_rng"]
+__all__ = [
+    "seed_rng",
+    "JOURNAL_PATH",
+    "DEFAULT_GRAPH",
+    "load_spikes",
+    "average_interval",
+    "load_journal",
+    "compute_trend",
+    "log_spike",
+    "log_score",
+    "capture_seed",
+    "seed_random",
+    "GraphBuilder",
+    "GraphIntegrator",
+]
